docker/docling-graph/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
docling-graph>=0.1.0
litellm>=1.0.0
networkx>=3.0
pyyaml>=6.0
httpx>=0.27.0
